pyproject.toml
[project]
name = "image-wizard"
version = "0.1.0"
description = "Local, on-device photo indexer with EXIF, YOLO, faces, and CLIP search."
requires-python = ">=3.12,<3.13"
dependencies = [
    # Core
    "typer>=0.12",
    "rich>=13.7",
    "platformdirs>=4.2",
    # Images
    "pillow>=10.3",
    "pillow-heif>=0.16",
    "pi-heif>=0.16",
    "numpy>=1.26,<2.3",
    # Metadata
    "pyexiftool>=0.5.6",
    # Geo
    "reverse-geocoder>=1.5.1",
    # Vector store
    "sqlite-vec>=0.1.6",
    # ML (CPU/MPS)
    "torch>=2.3",
    "torchvision>=0.18",
    "ultralytics>=8.3",
    "insightface>=0.7.3",
    "onnxruntime>=1.18",
    "open-clip-torch>=2.24",
    # Clustering
    "hdbscan>=0.8.38",
    "scikit-learn>=1.4",
    # Web
    "fastapi>=0.111",
    "uvicorn>=0.30",
    "jinja2>=3.1",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
raw = ["rawpy>=0.21"]
dev = [
    "pytest>=8.2",
    "pytest-cov>=5.0",
]

[project.scripts]
image-wizard = "imagewizard.cli:app"
iwiz = "imagewizard.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["imagewizard"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
